% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroCharms.R
\name{charmsToggle}
\alias{charmsToggle}
\title{Create a Metro 4 charms toggle}
\usage{
charmsToggle(..., id)
}
\arguments{
\item{...}{Button text or icon.}

\item{id}{Charms unique id. Needed by the charmsToggle.}
}
\description{
Build a Metro charms toggle button
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    fluidRow(
     charmsToggle(id = "mycharm", "Toggle charms")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
