% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroBadge.R
\name{metroBadge}
\alias{metroBadge}
\title{Create a Metro 4 badge}
\usage{
metroBadge(..., position = NULL, text_color = NULL, color = NULL)
}
\arguments{
\item{...}{Badge text.}

\item{position}{Badge position: inside, inline or NULL}

\item{text_color}{Badge text color. Black by default.}

\item{color}{Badge background color. No background by default.}
}
\description{
Build a Metro badge
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
     actionButton(inputId = "a", label = metroBadge(4)),
     actionButton(inputId = "b", label = metroBadge(4, position = "inline", color = "black")),
     actionButton(inputId = "c", label = metroBadge(4, position = "inside", color = "red", text_color = "white"))
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
