% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-grids.R
\name{metroCell}
\alias{metroCell}
\title{Create a Metro 4 cell}
\usage{
metroCell(..., width = NULL, offset = NULL, gap = TRUE)
}
\arguments{
\item{...}{Any element.}

\item{width}{Cell width. Between 1 and 12.}

\item{offset}{Numeric: cell offset. Between 1 and 12}

\item{gap}{Whether to display gaps between cells. TRUE by default.}
}
\description{
Build a Metro cell
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
