% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroCharms.R
\name{metroCharms}
\alias{metroCharms}
\title{Create a Metro 4 charms}
\usage{
metroCharms(
  ...,
  id,
  position = c("right", "left", "top", "bottom"),
  opacity = NULL,
  background = "back"
)
}
\arguments{
\item{...}{Any UI element.}

\item{id}{Charms unique id. Needed by the charmsToggle.}

\item{position}{Charms position: right, left, top or bottom.}

\item{opacity}{Charms opacity. NULL by default. Between 0 and 1.}

\item{background}{Charms background color. "black" by default.}
}
\description{
Build a Metro charms
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    br(), br(), br(), br(),
    charmsToggle(id = "mycharm", "Toggle charms"),
    metroCharms(
     id = "mycharm",
     background = "dark",
     position = "top",
     opacity = "0.5",
     sliderInput(
      "obs",
      "Number of observations:",
      min = 0,
      max = 1000,
      value = 500
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
