% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-inputs.R
\name{metroCheck}
\alias{metroCheck}
\title{Checkbox}
\usage{
metroCheck(
  inputId,
  label,
  value = FALSE,
  disabled = FALSE,
  style = c("1", "2"),
  labelSide = c("left", "right")
)
}
\arguments{
\item{inputId}{Unique input id.}

\item{label}{Checkbox label.}

\item{value}{Whether the checkbox is checked at start. Default to FALSE.}

\item{disabled}{Whether the checkbox is disabled at start. Default to FALSE.}

\item{style}{Checkbox style: either "1" (fill) or "3" (outline).}

\item{labelSide}{Label side: "left" or "right".}
}
\description{
Create a customizable checkbox with support for highly used options
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMetroUI)

 shiny::shinyApp(
  ui = metroPage(
    metroCheck(
     inputId = "checkbox1",
     label = "Checkbox",
     labelSide = "right"
   ),
    metroCheck(
     inputId = "checkbox2",
     label = "Checkbox checked",
     value = TRUE,
     labelSide = "right"
    ),
    metroCheck(
     inputId = "checkbox3",
     label = "Checkbox disabled",
     disabled = TRUE,
     labelSide = "right"
    ),
    metroCheck(
     inputId = "checkbox4",
     label = "Checkbox style 2",
     style = "2",
     labelSide = "left"
    ),
  ),
  server = function(input, output) {}
 )
}
}
