% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroCol.R
\name{metroCol}
\alias{metroCol}
\title{Create a Metro 4 column cell}
\usage{
metroCol(..., width, offset = NULL)
}
\arguments{
\item{...}{Any UI element.}

\item{width}{Column width. Between 1 and 12.}

\item{offset}{Column offset (shift). NULL by default.}
}
\description{
Build a Metro column
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    fluidRow(
     metroCol(width = 4, 1),
     metroCol(width = 4, 2),
     metroCol(width = 4, 3)
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
