% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-grids.R
\name{metroContainer}
\alias{metroContainer}
\title{Create a Metro 4 container}
\usage{
metroContainer(..., fullWidth = FALSE)
}
\arguments{
\item{...}{Any element.}

\item{fullWidth}{Whether the container should take the whole width. FALSE by default.}
}
\description{
Build a Metro container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
