% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-grids.R
\name{metroGrid}
\alias{metroGrid}
\title{Create a Metro 4 grid}
\usage{
metroGrid(...)
}
\arguments{
\item{...}{Any element.}
}
\description{
Build a Metro grid
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
