% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroPage.R
\name{metroPage}
\alias{metroPage}
\title{Create a Metro 4 page}
\usage{
metroPage(..., title = NULL)
}
\arguments{
\item{...}{Any element. They are inserted in a grid. Use the shiny fluidRow function
to create a row and insert metroUiCol inside. The maximum with is 12
(3 columns or lenght 4, 4 columns of lenght 3, ...).}

\item{title}{Page title.}
}
\description{
Build a metroUI page
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
