% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroPanel.R
\name{metroPanel}
\alias{metroPanel}
\title{Create a Metro 4 Panel}
\usage{
metroPanel(
  ...,
  title = NULL,
  width = NULL,
  height = NULL,
  collapsible = FALSE,
  collapsed = FALSE,
  draggable = FALSE,
  shadow = NULL,
  title_color = NULL,
  icon_color = NULL,
  collapse_color = NULL,
  content_color = NULL
)
}
\arguments{
\item{...}{Panel content.}

\item{title}{Panel title.}

\item{width}{Panel width.}

\item{height}{Panel height.}

\item{collapsible}{Whether the panel is collapsible or not. FALSE by default.}

\item{collapsed}{Whether the panel is collapsed at start. FALSE by default.}

\item{draggable}{Whether the panel is draggable or not. FALSE by default.}

\item{shadow}{Panel z-index. Numeric value expected.}

\item{title_color}{Title background color.}

\item{icon_color}{Icon background color.}

\item{collapse_color}{Collapse icon background color.}

\item{content_color}{Content background color.}
}
\description{
Build a Metro panel
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    metroCol(
     metroPanel(
      title = "Panel title",
      width = "250",
      collapsible = TRUE,
      draggable = TRUE,
      "Raptus capios ducunt ad genetrix.
      Joy doesn’t beautifully respect any believer —
      but the power is what flies",
      shadow = 3,
      title_color = "red",
      icon_color = "green",
      collapse_color = "dark",
      content_color = "cyan"
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
