% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-grids.R
\name{metroRow}
\alias{metroRow}
\title{Create a Metro 4 row}
\usage{
metroRow(..., xAlign = c("start", "center", "end"), yAlign = c("start",
  "center", "end"))
}
\arguments{
\item{...}{Any element.}

\item{xAlign}{How to horizontally align elements in the row: "start", "center" or "end".}

\item{yAlign}{How to vertically align elements in the row: "start", "center" or "end".}
}
\description{
Build a Metro row
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
