% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-grids.R
\name{metroSplitter}
\alias{metroSplitter}
\title{Splitter}
\usage{
metroSplitter(panel1, panel2, vertical = FALSE, minSize = 100, gutterSize = 4)
}
\arguments{
\item{panel1}{Left panel. Wrap in \link[shiny]{tagList}. This can be a nested \link{metroSplitter}.}

\item{panel2}{Right panel. Wrap in \link[shiny]{tagList}. This can be a nested \link{metroSplitter}.}

\item{vertical}{Vertical split. FALSE by default.}

\item{minSize}{Minimum panel size. 100px by default.}

\item{gutterSize}{Minimum gutter size (separator). 4 px by default.}
}
\description{
Create a layout with panels that can be resized with splitter component
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMetroUI)
}
}
