% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-tabs.R
\name{metroTab}
\alias{metroTab}
\title{Create a Metro 4 tab}
\usage{
metroTab(..., tabName, disable = FALSE)
}
\arguments{
\item{...}{Tab content.}

\item{tabName}{Unique tab name.}

\item{disable}{Whether the tab is disabled. FALSE by default}
}
\description{
Build a Metro tab
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
