% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-tabs.R
\name{metroTabs}
\alias{metroTabs}
\title{Create a Metro 4 tab container}
\usage{
metroTabs(..., mode = NULL)
}
\arguments{
\item{...}{Slot for \link{metroTab}.}

\item{mode}{Tab mode: NULL or "pills"}
}
\description{
Build a Metro tab container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
     metroTabs(
      metroTab(tabName = "Tab 1", "test 1"),
      metroTab(tabName = "Tab 2", "test 2"),
      metroTab(tabName = "Tab 3", "test 3")
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
