% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-tiles.R
\name{metroTilesGrid}
\alias{metroTilesGrid}
\title{Create a Metro 4 Tiles Grid}
\usage{
metroTilesGrid(..., group = FALSE, title = NULL, size = 2)
}
\arguments{
\item{...}{Insert metroTile inside.}

\item{group}{Whether tiles are displayed by group. FALSE by default.}

\item{title}{If group is TRUE, the group title.}

\item{size}{Tile group size: between 1 and 10.}
}
\description{
Build a Metro grid for tiles
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    metroTilesGrid(
     metroTile(size = "small", color = "red"),
     metroTile(size = "small", color = "green"),
     metroTile(size = "small", color = "blue", col_position = 1, row_position = 2),
     metroTile(size = "small", color = "orange", col_position = 2, row_position = 2),
     metroTile(size = "wide", color = "brown"),
     metroTile(size = "medium", color = "green", selected = TRUE)
    ),
    br(), br(), br(),
    metroTilesGrid(
     group = TRUE,
     size = 2,
     metroTile(size = "small", color = "red"),
     metroTile(size = "small", color = "green"),
     metroTile(size = "small", color = "blue", col_position = 1, row_position = 2),
     metroTile(size = "small", color = "orange", col_position = 2, row_position = 2),
     metroTile(
      size = "wide",
      color = "pink",
      sliderInput("obs", "Number of observations:",
       min = 0, max = 1000, value = 500
      )
     ),
     metroTile(
      size = "large",
      color = "green",
      selected = TRUE,
      plotOutput("distPlot")
     )
    )
   ),
   server = function(input, output) {
    output$distPlot <- renderPlot({
     hist(rnorm(input$obs))
    })
   }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
