% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroUiContainer.R
\name{metroUiContainer}
\alias{metroUiContainer}
\title{Create a Metro 4 container}
\usage{
metroUiContainer(...)
}
\arguments{
\item{...}{Any element.}
}
\description{
Build a metroUI container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroUiPage(
    metroUiContainer()
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
