% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroUiPage.R
\name{metroUiPage}
\alias{metroUiPage}
\title{Create a Metro 4 page}
\usage{
metroUiPage(..., title = NULL)
}
\arguments{
\item{...}{Any element.}

\item{title}{Page title.}
}
\description{
Build a metroUI page
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroUiPage(),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
