% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroWindow.R
\name{metroWindow}
\alias{metroWindow}
\title{Create a Metro 4 window}
\usage{
metroWindow(
  ...,
  title = NULL,
  icon = NULL,
  shadow = TRUE,
  header_color = NULL,
  content_color = NULL,
  content_text_color = NULL,
  minimize = TRUE,
  closable = TRUE,
  draggable = FALSE,
  resizable = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{...}{Any UI element.}

\item{title}{Window title.}

\item{icon}{Window icon.}

\item{shadow}{Whether to display a shadow around the window. TRUE by default.}

\item{header_color}{Header background color.}

\item{content_color}{Content background color.}

\item{content_text_color}{Content text color.}

\item{minimize}{Whether the window can be minimized. TRUE by default.}

\item{closable}{Whether the window can be closed. TRUE by default.}

\item{draggable}{Whether the window can be dragged. FALSE by default.}

\item{resizable}{Whether the window can be resized. FALSE by default.}

\item{width}{Window width.}

\item{height}{Window height.}
}
\description{
Build a Metro window
}
\note{
It is better to use only 2 windows per page and embed them as in the
example below.
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
  ui = metroPage(
    allow_loading = TRUE,
    fluidRow(
      metroCell(
        metroWindow(
          title = "Window",
          resizable = TRUE,
          width = "100\%",
          header_color = "red",
          content_color = "black",
          content_text_color = "yellow",
          # content
          metroTilesGrid(
            metroTile(
              size = "small",
              color = "indigo",
              icon = "github",
              url = "https://github.com/olton/Metro-UI-CSS"),
            metroTile(size = "small", color = "green", icon = "envelop"),
            metroTile(size = "small", color = "blue", col_position = 1, row_position = 2),
            metroTile(size = "small", color = "orange", col_position = 2, row_position = 2),
            metroTile(
              size = "wide",
              color = "pink",
              sliderInput("obs", "Number of observations:",
                          min = 0, max = 1000, value = 500
              )
            ),
            metroTile(
              size = "large",
              color = "green",
              selected = TRUE,
              plotOutput("distPlot")
            )
          )
        )
      ),
      metroCell(
        metroWindow(
          title = "Window",
          resizable = TRUE,
          width = "100\%",
          header_color = "red",
          content_color = "black",
          content_text_color = "yellow",
          # content
          metroTilesGrid(
            metroTile(size = "large",color = "brown", icon = "windows"),
            metroTile(size = "wide", color = "crimson", icon = "envelop"),
            metroTile(size = "small", color = "cobalt", col_position = 1, row_position = 2),
            metroTile(size = "small", color = "olive", col_position = 2, row_position = 2)
          )
        )
      )
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      hist(rnorm(input$obs))
    })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
