% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro-update-inputs.R
\name{updateMetroCheck}
\alias{updateMetroCheck}
\title{Change the value of a checkbox input on the client}
\usage{
updateMetroCheck(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to the server.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}
}
\description{
Change the value of a checkbox input on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyMetroUI)

 shiny::shinyApp(
  ui = metroPage(
    actionButton("go", "Update checkbox"),
    metroCheck(inputId = "checkbox1", label = "Checkbox", labelSide = "right")
  ),
  server = function(input, output, session) {
    observeEvent(input$go, {
      updateMetroCheck(session, inputId = "checkbox1", value = TRUE)
    })
  }
 )
}
}
