% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroBadge.R
\name{metroBadge}
\alias{metroBadge}
\title{Create a Metro 4 badge}
\usage{
metroBadge(..., position = NULL, text_color = NULL, color = NULL)
}
\arguments{
\item{...}{Badge text.}

\item{position}{Badge position: inside, inline or NULL}

\item{text_color}{Badge text color. Black by default.}

\item{color}{Badge background color. No background by default.}
}
\description{
Build a Metro badge
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
     metroGrid(
      metroRow(
       xAlign = "center",
       yAlign = "center",
       metroCell(metroBadge(4, color = "red")),
       metroCell(metroBadge(4, color = "yellow")),
       metroCell(metroBadge(4, color = "green"))
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
