% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroContainer.R
\name{metroContainer}
\alias{metroContainer}
\title{Create a Metro 4 container}
\usage{
metroContainer(...)
}
\arguments{
\item{...}{Any element.}
}
\description{
Build a Metro container
}
\examples{
if(interactive()){
 library(shiny)
 library(shinyMetroUi)

 shiny::shinyApp(
   ui = metroPage(
    metroContainer()
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
